requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
python-dotenv==1.0.0
httpx[http2]==0.27.2
sqlalchemy
psycopg2